added the validation for no draw at the knockout level
