+++ internal/handlers/championships.go
@@ -606,6 +606,7 @@ func updateChampionshipAfterMatch(ctx context.Context, championshipID, fixtureID
 	team1Score := 0
 	team2Score := 0
 	var winnerID *primitive.ObjectID
+	isDraw := false
 
 	// Prefer teamStats if present (legacy), otherwise use teamA/teamB
 	if teamStatsVal, ok := data["teamStats"]; ok && teamStatsVal != nil {
@@ -645,16 +646,25 @@ func updateChampionshipAfterMatch(ctx context.Context, championshipID, fixtureID
 		winnerID = &fixture.Team1ID
 	} else if team2Score > team1Score && fixture.Team2ID != nil {
 		winnerID = fixture.Team2ID
+	} else {
+		isDraw = true
 	}
-	// Note: draws shouldn't happen in knockout, but if they do, winnerID remains nil
 
-	// Determine winner
-	if team1Score > team2Score {
-		winnerID = &fixture.Team1ID
-	} else if team2Score > team1Score && fixture.Team2ID != nil {
-		winnerID = fixture.Team2ID
+	if isDraw {
+		var championship models.Championship
+		err = db.ChampionshipsCollection.FindOne(
+			ctx,
+			bson.M{"_id": fixture.ChampionshipID},
+		).Decode(&championship)
+		if err != nil {
+			return err
+		}
+
+		isSemifinalOrFinal := fixture.RoundNumber >= championship.TotalRounds-1
+		if isSemifinalOrFinal {
+			return fmt.Errorf("%w: championship semifinal/final match cannot end in a tie", ErrKnockoutTieNotAllowed)
+		}
 	}
-	// Note: draws shouldn't happen in knockout, but if they do, winnerID remains nil
 
 	// Update fixture with scores and winner
 	updateDoc := bson.M{
